Fix lint: split test helpers by build tag; fix SSRF: use switch for HEAD URL
- Move integration-only test helpers to test_helpers_integration_test.go
  with //go:build integration tag so golangci-lint no longer flags them as
  unused when run without -tags integration
- Keep setupTestDB in test_helpers_test.go (used by non-integration security tests)
- Fix SSRF (CodeQL alert #40): construct HEAD request URL via switch on
  parsed.Host so scheme+host are server-controlled constants, not derived
  from user input
